requirements.txt
instagrapi
openai>=1.0.0
python-dotenv
Pillow
pydantic
